lint
SCCS-vsn: sys/hp300/dev/st.c 8.3
SCCS-vsn: sys/hp/dev/grf.c 8.4
SCCS-vsn: sys/hp/dev/ite_subr.c 8.2
SCCS-vsn: sys/hp300/hp300/kgdb_stub.c 8.4

+++ usr/src/sys/hp300/hp300/kgdb_stub.c
@@ -13,7 +13,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)kgdb_stub.c	8.3 (Berkeley) %G%
+ *	@(#)kgdb_stub.c	8.4 (Berkeley) %G%
  */
 
 /*
@@ -321,7 +321,7 @@ kgdb_trap(type, frame)
 	int inlen;
 	u_long gdb_regs[NUM_REGS];
 
-	if (kgdb_dev < 0) {
+	if ((int)kgdb_dev < 0) {
 		/* not debugging */
 		return (0);
 	}
